Beginning of calendar
Went out and forgot to bring a cable for my
pi zero so didn't feel like working on the clock
Instead began work on a different utility displa
I'd like to have, a calendar.

Eventually I'd like to populate the calendar from
my own google calendar but for the time being I'll
settle for getting simply highlighting the current
day of the month functional.