Adjusted analog face and time display.
It's really hard to beat just testing on the desired display
itself, just so you can see exactly what you do and don't like.

This time around I adjusted the thickness and color of the clock
face border, because it looks much nicer on the screen than in
the preview PNGs I generated.

I also added a second hand display, which is debatably useful on a
screen that's only going to refresh every few minutes, but it
makes the clock look that much more authentic so I feel it was
worth the effort. Perhaps I can adjust the script so it adds a
random delay to the second hand for variation, otherwise it should
display at about the same second every time if the cron job runs
every X minutes.