Simple date display, unfinished
Shifted the clock face over for the PHAT analog display, now displaying 
some date abbreviations to the left of the face.

Made the second hand randomly placed as the display isn't going to be 
updated with any frequency to where it will be accurate.

Still to possibly include:
Some indication of weather, lunar phase, etc.