Initial API work
This API code isn't functional yet, I
need to learn a bit more about how to properly use
the requests library to submit the API call to darksky
and fix the API usage here.

The point of the darksky script is to essentially call
and cache the response data for use by potentially
multiple utilities, so if I to add a low-refresh
weather display, or weather to the calendar, or
whatever in the future they can all draw from the
cached response data that can run at a different
frequency to keep API calls more regular and controlled.